backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
elasticsearch==8.11.0
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2
python-multipart==0.0.6
websockets==12.0

